[test] see if routes are working
