+++ database/models/exports.js
@@ -1,0 +1,11 @@
+module.exports = {
+  Type: require('./Type'),
+  User: require('./User'),
+  UserLike: require('./user_like'),
+  Event: require('./Event'),
+  EventSchedule: require('./event_schedule'),
+  Schedule: require('./Schedule'),
+  Photo: require('./Photo'),
+  Hometown: require('./Hometown'),
+  UserHometown: require('./user_hometown'),
+};
