+++ database/models/user_like.js
@@ -1,0 +1,9 @@
+const Sequelize = require('sequelize');
+const sequelize = require('../database');
+
+// define user_like
+const UserLike = sequelize.define('user_like', {
+  like: Sequelize.BOOLEAN
+});
+
+module.exports = UserLike;
